Move io functions from _misc to io
