+++ datacube/utils/__init__.py
@@ -6,6 +6,7 @@ from .dates import datetime_to_seconds_since_1970, parse_time
 from .py import cached_property, ignore_exceptions_if, import_function
 from .serialise import jsonify_document
 from .uris import is_url, uri_to_local_path, get_part_from_uri, mk_part_uri
+from .io import slurp, check_write_path, write_user_secret_file
 from .documents import (
     InvalidDocException,
     SimpleDocNav,
@@ -29,11 +30,10 @@ from .math import (
 from ._misc import (
     DatacubeException,
     schema_validated,
-    write_user_secret_file,
-    slurp,
     gen_password,
     _readable_offset,
 )
+
 
 __all__ = (
     "datetime_to_seconds_since_1970",
@@ -66,6 +66,7 @@ __all__ = (
     "schema_validated",
     "write_user_secret_file",
     "slurp",
+    "check_write_path",
     "gen_password",
     "_readable_offset",
 )
